just commenting on functions on what to do, its too late and I need to sleep, dont want to make any more mistakes

+++ front-end/src/app/home/home.component.ts
@@ -117,6 +117,13 @@ export class HomeComponent implements OnInit {
   }
 
   deleteGroup(groupName){
+    if (!(this.user_obj.superadmin || this.user_obj.group_admin))
+    {
+      alert("No permissions");
+      return;
+    }
+
+
     let data = {name:groupName, user:this.username}
     this._groupService.deleteGroup(data).subscribe(
       data=>{
